Don't assume qt_tablet_target widget exists from one event to the next
In this scenario there is a widget with a button inside, and you click
the button with the tablet.  The target of the event is the button,
but when you click it, the parent (or ancestor) is destroyed.  Commit
2bac49265efcf8faabc3756d1a3e405a3d336f68 took care of the case when
the parent is a window, but it is not always a window which is being
destroyed.  So the approach of using a QPointer is better because it
should take care of all the cases when the qt_tablet_target is
destroyed during the course of a tablet event sequence.

Task-number: QTBUG-36848
Task-number: QTBUG-38040
Change-Id: Ia0e861f2cb2fbc30234aa596f3a36ddd0835a9af
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>