qdoc: Links to Q_DECL_FINAL not created
This update fixes part of QTBUG-37719. Q_DECL_FINAL is a
macro that is documented, but links to the documentation
for it were not created. The problem was that Q_DECL_FINAL
is actually used in the Qt sources, but it was not listed
in the Cpp.ignoretokens clause in the qdocconf file. This
update just adds it to Cpp.ignoretokens.

Task-number: QTBUG-37719
Change-Id: I62945df4976b21b67084e6de25e25bbc4c92f89f
Reviewed-by: Sze Howe Koh <szehowe.koh@gmail.com>
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>